separated actions and reducers
