+++ react/app.js
@@ -2,17 +2,17 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 import {createStore} from 'redux';
-import Reducers from 'reducers';
-import PublishTool from 'containers/PublishTool';
+import Reducer from 'reducers';
+import Publish from 'containers/PublishTool';
 
 let store = createStore(
-  Reducers,
+  Reducer,
   window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
 );
 
 ReactDOM.render(
   <Provider store={store}>
-    <PublishTool />
+    <Publish />
   </Provider>,
   document.getElementById('react-publish-tool')
 )
