+++ react/constants/action_types.js
@@ -1,9 +1,11 @@
-// export action types
+// channel actions
+export const CHANNEL_UPDATE = 'CHANNEL_UPDATE';
+
+// publish actions
 export const FILE_SELECTED = 'FILE_SELECTED';
 export const FILE_CLEAR = 'FILE_CLEAR';
 export const METADATA_UPDATE = 'METADATA_UPDATE';
 export const CLAIM_UPDATE = 'CLAIM_UPDATE';
-export const CHANNEL_UPDATE = 'CHANNEL_UPDATE';
 export const SET_PUBLISH_IN_CHANNEL = 'SET_PUBLISH_IN_CHANNEL';
 export const PUBLISH_STATUS_UPDATE = 'PUBLISH_STATUS_UPDATE';
 export const ERROR_UPDATE = 'ERROR_UPDATE';
