+++ react/containers/PublishMetadataInputs/index.js
@@ -1,12 +1,12 @@
 import {connect} from 'react-redux';
-import {updateMetadata} from 'actions';
+import {updateMetadata} from 'actions/publish';
 import View from './view';
 
-const mapStateToProps = state => {
+const mapStateToProps = ({ publish }) => {
   return {
-    description: state.metadata.description,
-    license    : state.metadata.license,
-    nsfw       : state.metadata.nsfw,
+    description: publish.metadata.description,
+    license    : publish.metadata.license,
+    nsfw       : publish.metadata.nsfw,
   };
 };
